Refactored the same codes
