+++ embedded/src/integTest/groovy/com/bmuschko/gradle/tomcat/embedded/fixture/EmbeddedTomcat7xPlusIntegrationTest.groovy
@@ -1,0 +1,9 @@
+package com.bmuschko.gradle.tomcat.embedded.fixture
+
+abstract class EmbeddedTomcat7xPlusIntegrationTest extends EmbeddedTomcatIntegrationTest {
+    @Override
+    protected void configureTomcatServer() {
+        tomcatServer.embedded.getHost()
+        tomcatServer.embedded.port = port
+    }
+}
