+++ tomcat7x/src/integTest/groovy/com/bmuschko/gradle/tomcat/embedded/Tomcat7xServerIntegrationTest.groovy
@@ -15,32 +15,16 @@
  */
 package com.bmuschko.gradle.tomcat.embedded
 
-import com.bmuschko.gradle.tomcat.embedded.fixture.EmbeddedTomcatIntegrationTest
+import com.bmuschko.gradle.tomcat.embedded.fixture.EmbeddedTomcat7xPlusIntegrationTest
 
 /**
  * Tomcat 7x server test.
  *
  * @author Benjamin Muschko
  */
-class Tomcat7xServerIntegrationTest extends EmbeddedTomcatIntegrationTest {
+class Tomcat7xServerIntegrationTest extends EmbeddedTomcat7xPlusIntegrationTest {
     @Override
     protected TomcatServer createTomcatServer() {
         new Tomcat7xServer()
-    }
-
-    @Override
-    protected File getTomcatHomeDir() {
-        temporaryFolder.newFolder('tomcat7xHome')
-    }
-
-    @Override
-    protected void configureTomcatServer() {
-        tomcatServer.embedded.getHost()
-        tomcatServer.embedded.port = port
-    }
-
-    @Override
-    protected TomcatVersion getTomcatVersion() {
-        TomcatVersion.VERSION_7_0_X
     }
 }
